requirements.txt
altgraph==0.17.2
auto-py-to-exe==2.23.0
bottle==0.12.23
bottle-websocket==0.2.9
cffi==1.15.1
Eel==0.14.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.3
numpy==1.22.3
pefile==2022.5.30
Pillow==9.2.0
pycparser==2.21
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pyparsing==3.0.9
pywin32-ctypes==0.2.0
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
